# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "News Bots V2 API",
    "description": "API for managing All routes related to News Bots",
    "version": "2.0"
  },
...
  "paths": {
...
    "/articles/create": {
      "post": {
        "tags": [
          "News Creator"
        ],
        "summary": "Create a new article",
        "description": "Create a new article with the provided data.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content",
                  "analysis",
                  "used_keywords",
                  "is_article_efficient",
                  "bot_id",
                  "category_id",
                  "image_url"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "analysis": {
                    "type": "string"
                  },
                  "used_keywords": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "is_article_efficient": {
                    "type": "string"
                  },
                  "bot_id": {
                    "type": "integer"
                  },
                  "category_id": {
                    "type": "integer"
                  },
                  "image_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article created successfully"
          },
          "400": {
            "description": "Bad request. Missing required fields"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/articles/generate_image": {
      "post": {
        "tags": [
          "News Creator"
        ],
        "summary": "Generate an image using DALL-E 3 model",
        "description": "Generate an image based on the content and bot_id using DALL-E 3 model.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "bot_id"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "bot_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image generated successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/articles/generate": {
      "post": {
        "tags": [
          "News Creator"
        ],
        "summary": "Generate an article summary using the Perplexity model",
        "description": "Generate a new article summary based on the provided content and category_id.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "category_id"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "category_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article summary generated successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/articles/recent": {
      "get": {
        "tags": [
          "News Creator"
        ],
        "summary": "Get the last five articles",
        "description": "Retrieve the last five articles with extended details.",
        "responses": {
          "200": {
            "description": "Successfully retrieved the last five articles"
          },
          "404": {
            "description": "No articles found matching the criteria"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/content/extract": {
      "post": {
        "tags": [
          "News Creator"
        ],
        "summary": "Extract content from a link or Google Docs",
        "description": "Extract content from a provided link or Google Docs based on the extract type.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "extract_type",
                  "link"
                ],
                "properties": {
                  "extract_type": {
                    "type": "string",
                    "description": "Type of extraction ('link' or 'google_docs')"
                  },
                  "link": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully extracted content"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/upload_news_file_to_drive": {
      "post": {
        "tags": [
          "News Creator"
        ],
        "summary": "Upload a file to Google Drive",
        "description": "Upload a file to a specified folder in Google Drive.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "folderName",
                  "file",
                  "fileName"
                ],
                "properties": {
                  "folderName": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "fileName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          },
          "400": {
            "description": "Bad request. Missing required data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}